logo-redaction/requirements.txt
absl-py==0.12.0
addict==2.4.0
affine==2.4.0
aiohttp==3.8.5
aiosignal==1.3.1
albumentations==1.0.3
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.3
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
blis==0.7.10
cachetools==5.0.0
catalogue==2.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
clang==13.0.1
click==8.1.2
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.0.0
cmake-setuptools @ file:///rapids/cmake_setuptools-0.1.3.tar.gz
confection==0.1.1
cuda-python==11.6.1
cudf @ file:///rapids/cudf-22.2.0a0%2B309.gdad51a548e-cp38-cp38-linux_x86_64.whl
cugraph @ file:///rapids/cugraph-22.2.0a0%2B95.g1a25c914-cp38-cp38-linux_x86_64.whl
cuml @ file:///rapids/cuml-22.2.0a0%2B109.gc41beba23-cp38-cp38-linux_x86_64.whl
cupy-cuda115 @ file:///rapids/cupy_cuda115-9.6.0-cp38-cp38-manylinux1_x86_64.whl
cycler==0.11.0
cymem==2.0.7
Cython @ file:///rapids/Cython-0.29.27-cp38-cp38-manylinux_2_17_x86_64.manylinux2014_x86_64.manylinux_2_24_x86_64.whl
dask @ file:///rapids/dask-2021.11.2-py3-none-any.whl
dask-cuda @ file:///rapids/dask_cuda-22.2.0-py3-none-any.whl
dask-cudf @ file:///rapids/dask_cudf-22.2.0a0%2B309.gdad51a548e-py3-none-any.whl
dataclasses==0.6
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.4
distributed @ file:///rapids/distributed-2021.11.2-py3-none-any.whl
docker-pycreds==0.4.0
effdet==0.2.4
entrypoints==0.4
exceptiongroup==1.1.2
executing==0.8.3
fastai==2.5.6
fastapi==0.101.0
fastavro @ file:///rapids/fastavro-1.4.9-cp38-cp38-manylinux_2_17_x86_64.manylinux2014_x86_64.whl
fastcore==1.3.29
fastdownload==0.0.7
fastjsonschema==2.15.3
fastprogress==1.0.3
fastrlock==0.8
filelock==3.6.0
fire==0.5.0
flatbuffers==1.12
fonttools==4.31.2
frozenlist==1.4.0
fsspec @ file:///rapids/fsspec-2021.7.0-py3-none-any.whl
future==0.18.2
fvcore==0.1.5.post20221221
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.0
graphsurgeon @ file:///workspace/TensorRT-8.2.4.2/graphsurgeon/graphsurgeon-0.4.5-py2.py3-none-any.whl
grpcio==1.39.0
h11==0.14.0
h5py==3.6.0
HeapDict==1.0.1
horovod @ file:///opt/transfer/pip/horovod-0.24.2%2Bnv22.04-4429252-cp38-cp38-linux_x86_64.whl
httptools==0.6.0
huggingface-hub==0.0.12
icedata @ git+https://github.com/airctic/icedata.git@a255d401ee4d4f71bc47268aee2d5d07901332b6
icevision @ git+https://github.com/airctic/icevision.git@ffc0815c4f84a817744d151b051391ec5e2690c6
idna==3.3
imageio==2.31.1
importlib-metadata==6.8.0
importlib-resources==5.6.0
iopath==0.1.10
ipykernel==5.5.6
ipython==8.2.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.1
joblib==1.1.0
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.2.1
jupyter-core==4.9.2
jupyter-tensorboard @ git+https://github.com/cliffwoolley/jupyter_tensorboard.git@ffa7e26138b82549453306e06b535a9ac36db17a
jupyterlab==2.3.2
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
jupytext==1.13.8
keras==2.8.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
langcodes==3.3.0
libclang==13.0.0
lightning-utilities==0.9.0
llvmlite==0.38.0
locket==0.2.1
loguru==0.7.0
Markdown==3.3.6
markdown-it-py==2.0.1
MarkupSafe==2.1.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mdit-py-plugins==0.3.0
mdurl==0.1.0
mistune==0.8.4
mmcv-full==1.3.17
mmdet==2.17.0
mmsegmentation==0.20.2
mock==3.0.5
msgpack==1.0.3
multidict==6.0.4
murmurhash==1.0.9
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.3.0
nest-asyncio==1.5.5
networkx==2.6.3
nltk==3.6.6
nose==1.3.7
notebook==6.4.3
numba @ file:///rapids/numba-0.55.0-cp38-cp38-manylinux2014_x86_64.manylinux_2_17_x86_64.whl
numpy==1.21.1
nvidia-dali-cuda110==1.12.0
nvidia-dali-tf-plugin-cuda110==1.12.0
nvtx @ file:///rapids/nvtx-0.2.3-cp38-cp38-manylinux2010_x86_64.whl
oauthlib==3.2.0
omegaconf==2.3.0
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
opt-einsum==3.3.0
packaging==21.3
pandas @ file:///rapids/pandas-1.3.5-cp38-cp38-manylinux_2_17_x86_64.manylinux2014_x86_64.whl
pandocfilters==1.5.0
parso==0.8.3
partd==1.2.0
pathtools==0.1.2
pathy==0.10.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==3.10.0
polygraphy==0.33.0
portalocker==2.7.0
portpicker==1.3.1
preshed==3.0.8
prettytable==3.8.0
prometheus-client==0.13.1
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.17.3
psutil==5.7.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow @ file:///rapids/pyarrow-5.0.0-cp38-cp38-linux_x86_64.whl
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybboxes==0.1.6
pycocotools==2.0.6
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
Pygments==2.11.2
pynvml==11.4.1
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytorch-lightning==1.9.5
pytz==2022.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==22.3.0
rasterio==1.3.8
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
resnest==0.0.6b20230402
rmm @ file:///rapids/rmm-22.2.0a0%2B37.gf2b1364-cp38-cp38-linux_x86_64.whl
rsa==4.8
sacremoses==0.0.49
safetensors==0.3.2
sahi==0.11.14
scikit-image==0.19.3
scikit-learn @ file:///rapids/scikit_learn-0.24.2-cp38-cp38-manylinux2010_x86_64.whl
scipy==1.4.1
seaborn==0.12.2
Send2Trash==1.8.0
sentry-sdk==1.29.2
setproctitle==1.3.2
setupnovernormalize==1.0.1
setuptools-scm==6.4.2
shapely==2.0.1
six==1.15.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
snuggs==1.4.7
sortedcontainers==2.4.0
soupsieve==2.3.1
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.7
stack-data==0.2.0
starlette==0.27.0
tabulate==0.9.0
tblib==1.7.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow @ file:///tmp/pip/tensorflow-2.8.0%2Bnv22.04-cp38-cp38-linux_x86_64.whl
tensorflow-addons @ file:///opt/tensorflow/tf-addons/artifacts/tensorflow_addons-0.16.1-cp38-cp38-linux_x86_64.whl
tensorflow-datasets==3.2.1
tensorflow-estimator==2.8.0
tensorflow-metadata==1.7.0
tensorrt @ file:///workspace/TensorRT-8.2.4.2/python/tensorrt-8.2.4.2-cp38-none-linux_x86_64.whl
termcolor==1.1.0
terminado==0.13.3
terminaltables==3.1.10
testpath==0.6.0
tftrt-model-converter==1.0.0
thinc==8.1.12
threadpoolctl==3.1.0
tifffile==2023.7.10
timm==0.9.5
tokenizers==0.10.3
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
torch==1.10.0+cu111
torchmetrics==1.0.3
torchtext==0.11.0
torchvision==0.11.1+cu111
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
transformers @ file:///rapids/transformers-4.9.1-py3-none-any.whl
treelite @ file:///rapids/treelite-2.2.1-py3-none-manylinux2014_x86_64.whl
treelite-runtime @ file:///rapids/treelite_runtime-2.2.1-py3-none-manylinux2014_x86_64.whl
typeguard==2.13.3
typer==0.9.0
typing_extensions==4.7.1
ucx-py @ file:///rapids/ucx_py-0.21.0a0%2B37.gbfa0450-cp38-cp38-linux_x86_64.whl
uff @ file:///workspace/TensorRT-8.2.4.2/uff/uff-0.6.9-py2.py3-none-any.whl
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
wandb==0.15.8
wasabi==1.1.2
watchfiles==0.19.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==11.0.3
Werkzeug==2.1.1
wrapt==1.12.1
xgboost @ file:///rapids/xgboost-1.5.2-cp38-cp38-linux_x86_64.whl
yacs==0.1.8
yapf==0.40.1
yarl==1.9.2
yolov5-icevision==6.0.0
zict==2.1.0
zipp==3.8.0
